backend/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
supabase==2.0.2
websockets>=14.0
pandas>=2.2.2
pytest>=8.2.0
pybaseball>=2.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
confluent-kafka>=2.3.0